Added Python versions for 'emb1' contributed by R.Wobst
The Python versions are compatible with the Tcl at the protocol level.

I.e. they allow to be run in any mix, e.g.
- Tcl server with Python client,
- Python server with one Python and two Tcl clients

Furthermore, the Python client is implemented twice, one uses TkInter, the other wxWidgets.

NOTE: Trace-Output is ONLY to the console and cannot (yet) be displayed with the Tcl-based
network viewers.