Provided a possibility to close tabs in the trace-view.
Trace-view tabs stay open even after the clients are stopped, filling it with its messages. This is because you
may want to see the final lines of output. (Especially in case of stalled or crashed programs these may give
the essential hints.)

When tracing many clients most of which "come and go", tabs collect though there is not any longer any added.
By removing any output from a tab for a terminated connection, now also the tab is removed.