Improved "Step by Step" description for trying out.
The improvements were based to the most part on the feedback by R. Wobst, who
recently started to contribute Python versions of the `emb1` example (see
directory `emb_py`).

(Hopefully) the wording is better now and the "step by step" style clearer,
especially when it comes to `emb2`. (Simulating the behavior of a real driver
with named pipes can become a little hairy, as it is essential to follow
the steps exactly, due to the rendezvous principle for opening such files.)

Furthermore the markdown to indicate examples and code was adapted, as it seems
not all formatters (e.g. like MD.py) recognize code blocks when enclosed in
lines of three back-ticks but require indentation by four spaces instead.